requirements.txt
# ============================================
# KAFKA
# ============================================
kafka-python==2.0.2
confluent-kafka==2.3.0

# ============================================
# SPARK
# ============================================
pyspark==3.5.0

# ============================================
# AIRFLOW
# ============================================
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-postgres==5.7.1

# ============================================
# DATABASE
# ============================================
psycopg2-binary==2.9.9
SQLAlchemy==1.4.50

# ============================================
# DATA PROCESSING
# ============================================
pandas==2.1.3
numpy==1.26.2

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# ============================================
# LOGGING & MONITORING
# ============================================
colorlog==6.8.0
python-json-logger==2.0.7

# ============================================
# TESTING (Optional but recommended)
# ============================================
pytest==7.4.3
pytest-cov==4.1.0

# ============================================
# DEVELOPMENT
# ============================================
black==23.12.0
flake8==6.1.0
autopep8==2.0.4